+++ scaffold/src/Service/ThemeStylesheetProvider.php
@@ -17,7 +17,7 @@ class ThemeStylesheetProvider implements BundleStylesheetProviderInterface
     private const string LAST_STYLESHEET = 'app.css';
 
     public function __construct(
-        #[Autowire('%kernel.project_dir%')]
+        #[Autowire(param: 'kernel.project_dir')]
         private readonly string $projectDir,
     ) {
     }