+++ src/Management/CollectionItemExportProvider.php
@@ -21,7 +21,7 @@ class CollectionItemExportProvider implements ExportProviderInterface
 {
     public function __construct(
         private readonly CollectionItemRepository $collectionItemRepository,
-        #[Autowire('%kernel.project_dir%')]
+        #[Autowire(param: 'kernel.project_dir')]
         private readonly string $projectDir,
     ) {
     }